+++ src/ui/section/components/OverviewZone/GenerateAndMore/GenerationButtons.h
@@ -1,8 +1,8 @@
 #pragma once
 
 #include <JuceHeader.h>
-#include "ui/extra/StyledButton.h"
-#include "ui/extra/ColoredPanel.h"
+#include "ui/extra/Button/StyledButton.h"
+#include "ui/extra/Component/Panel/ColoredPanel.h"
 
 //==============================================================================
 class GenerationButtons : public ColoredPanel
