+++ src/ui/section/components/OverviewZone/Overview/OverviewActionArea.h
@@ -1,7 +1,7 @@
 #pragma once
 
 #include <JuceHeader.h>
-#include "ui/extra/ColoredPanel.h"
+#include "ui/extra/Component/Panel/ColoredPanel.h"
 #include "../PlayBack/PlaybackActionArea.h"
 #include "../GenerateAndMore/GenerationButtons.h"
 #include "OverviewArea.h"
